refactor(http): extract HTTP/1 protocol handling from SocketHTTPServer.c
Extract HTTP/1.1 protocol handling functions to a dedicated module:
- Created SocketHTTPServer-http1.c (~600 LOC)
- Created SocketHTTPServer-http1.h header
- Extracted h2c upgrade (RFC 9113 §3.2) logic
- Extracted HTTP/1 request dispatching
- Extracted streaming body processing

Functions extracted:
- server_handle_parsed_request() - Main HTTP/1 request dispatcher
- server_try_h2c_upgrade() - HTTP/2 upgrade validation and transition
- server_header_has_token_ci() - RFC 9110 §5.1.5 token parsing
- server_decode_http2_settings() - Base64url decode HTTP2-Settings
- should_copy_header_to_h2() - RFC 9113 §3.2.1 header filtering
- server_process_streaming_body() - Streaming body callback handler

Performance improvements:
- Use Headers_get_n() with STRLEN_LIT() for known headers
- Reduces runtime strlen() calls for "Upgrade", "Connection", "HTTP2-Settings"

All 127 tests pass with sanitizers enabled.

Closes #2628
